Added an in-depth description and background
I realised that my previous one-liner readme file was useless to pretty much anyone other than me (and therefore so was the repository), so I revised it. It now has a couple of fancy headings and detailed paragraphs. This is an attempt to make my "portfolio" a little easier to understand for potential employers.